requirements.txt
fastapi[standard]
uvicorn
streamlit
firebase-admin
google-genai
moviepy
requests
beautifulsoup4
aiohttp
python-dotenv
